Move joi from dev dependencies to prod
Joi is used by this library and it was accidentally moved into
development dependencies ages ago. This change made it into v4.1.0.

I spotted this when I tried to upgrade next-control-centre to use the
latest version of this library - the app won't start because Joi is
missing as a dependency. This resolves the issue.